src/main/java/it/sbcourse/sbpostit/postit/entity/PostIt.java
package it.sbcourse.sbpostit.postit.entity;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Positive;
import jakarta.validation.constraints.Size;

public class PostIt {

    private static int counter;

    @NotNull
    @Positive
    private Integer id;

    @NotBlank
    @Size(min = 1, max = 10)
    private String msg;

    public PostIt() {
        this(null);
    }

    public PostIt(String msg) {
        this.id = ++counter;
        this.msg = msg;
    }

    public Integer getId() {
        return id;
    }

    public void setId(Integer id) {
        this.id = id;
    }

    public String getMsg() {
        return msg;
    }

    public void setMsg(String msg) {
        this.msg = msg;
    }

    @Override
    public int hashCode() {
        final int prime = 31;
        int result = 1;
        result = prime * result + ((id == null) ? 0 : id.hashCode());
        return result;
    }

    @Override
    public boolean equals(Object obj) {
        if (this == obj)
            return true;
        if (obj == null)
            return false;
        if (getClass() != obj.getClass())
            return false;
        PostIt other = (PostIt) obj;
        if (id == null) {
            if (other.id != null)
                return false;
        } else if (!id.equals(other.id))
            return false;
        return true;
    }

    @Override
    public String toString() {
        return "PostIt [id=" + id + ", msg=" + msg + "]";
    }

}
